+++ providers/plugins.rb
@@ -39,14 +39,14 @@ action :create do
 
   case @install_type
   when 'native'
-    ex = execute "bin/plugin install #{ls_name}" do
-      command "bin/plugin install #{ls_name}"
+    ex = execute "bin/logstash-plugin install #{ls_name}" do
+      command "bin/logstash-plugin install #{ls_name}"
       user    ls_user
       group   ls_group
       cwd     ls_instance_dir
       notifies :restart, "logstash_service[#{ls_instance}]"
       # this is a temp workaround to make the plugin command idempotent.
-      not_if { ::File.exist?("#{ls_instance_dir}/#{ls_install_check}") }
+      not_if "#{ls_instance_dir}/#{ls_install_check} #{ls_name}"
     end
     new_resource.updated_by_last_action(ex.updated_by_last_action?)
   when 'tarball'