+++ resources/instance.rb
@@ -25,10 +25,12 @@ attribute :user, kind_of: String
 attribute :group, kind_of: String
 attribute :create_account, kind_of: [TrueClass, FalseClass]
 attribute :join_groups, kind_of: [String]
-attribute :logrotate_enable, kind_of: [TrueClass, FalseClass]
+# attribute :logrotate_enable, kind_of: [TrueClass, FalseClass]
 attribute :user_opts, kind_of: [Hash]
-attribute :logrotate_size, kind_of: [String]
-attribute :logrotate_use_filesize, kind_of: [TrueClass, FalseClass]
-attribute :logrotate_frequency, kind_of: [String]
-attribute :logrotate_max_backup, kind_of: [Integer]
-attribute :logrotate_options, kind_of: [String]
+# attribute :logrotate_size, kind_of: [String]
+# attribute :logrotate_use_filesize, kind_of: [TrueClass, FalseClass]
+# attribute :logrotate_max_backup, kind_of: [Integer]
+# attribute :logrotate_options, kind_of: [String]
+attribute :log4j_frequency, kind_of: [Integer]
+attribute :log4j_template_cookbook, kind_of: [String]
+